Private API for iterating over QSvgTinyDocument
Expose more node information in the internal API. Also
use the new API to implement a debug operator for
QSvgTinyDocument. Includes a drive-by fix: use floating
point instead of int for rounded rect radius.

Task-number: QTBUG-116883

+++ src/svg/qsvggraphics.cpp
@@ -250,7 +250,7 @@ void QSvgPolyline::draw(QPainter *p, QSvgExtraStates &states)
     QT_SVG_TIMING_EXIT("Polyline")
 }
 
-QSvgRect::QSvgRect(QSvgNode *node, const QRectF &rect, int rx, int ry)
+QSvgRect::QSvgRect(QSvgNode *node, const QRectF &rect, qreal rx, qreal ry)
     : QSvgNode(node),
       m_rect(rect), m_rx(rx), m_ry(ry)
 {
